requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pydantic
google-generativeai==0.3.2
elevenlabs==0.2.27
pypdf
